Vendor mathsass instead of extending sass
When trying to integrate the latest changes into the main decidim repo,
I had trouble getting my custom sass extension to work with decidim, so
I tried another approach.

In any case, reimplementing this kind of math in ruby for frontend stuff
seemed wrong. I think we should:
* Reuse instead of reimplement.
* Use frontend tools (sass) instead of ruby.

So I decided to vendorize a sass library
(https://github.com/terkel/mathsass) that already implements these
calculations.

Another good thing about this is that "importing" (and automating the
import) from decidim-admin-design into decidim should be easier because
it's just a matter of copying the front end files and we don't need to
deal with any integrations of sass extensions or the like.